+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add Scan Local command to scan for devices on connected network
 - Respect Momentum settings for ESP UART Channel
 - Added PCAP and Wardrive clearing options in Settings menu
+- Added BLE Wardriving command
 - Miscellaneous bug fixes and improvements
 
 ## v1.1.7